refactor(repository-workflow): inline commit validation
背景：
- YAML 提交脚本已经负责提交前后的完整校验，保留独立 validator 会形成重复入口。

处理：
- 将 breaking change、Assisted-by 和 Git trailer 校验内聚到 commit_from_yaml.py。
- 删除 validate_commit.py，并把原有回归用例合并到 YAML 提交脚本测试中。

验证：
- 20 个 repository-workflow 测试全部通过。
- Ruff 和 skill 结构校验通过。
